+++ sdk/alertsmanagement/azure-mgmt-alertsmanagement/azure/mgmt/alertsmanagement/operations/_smart_detector_alert_rules_operations.py
@@ -18,11 +18,13 @@ from .. import models
 class SmartDetectorAlertRulesOperations(object):
     """SmartDetectorAlertRulesOperations operations.
 
+    You should not instantiate directly this class, but create a Client instance that will create it for you and attach it as attribute.
+
     :param client: Client for service requests.
     :param config: Configuration of service client.
     :param serializer: An object model serializer.
     :param deserializer: An object model deserializer.
-    :ivar api_version: Client Api Version. Constant value: "2019-03-01".
+    :ivar api_version: Client Api Version. Constant value: "2019-06-01".
     """
 
     models = models
@@ -32,15 +34,18 @@ class SmartDetectorAlertRulesOperations(object):
         self._client = client
         self._serialize = serializer
         self._deserialize = deserializer
-        self.api_version = "2019-03-01"
+        self.api_version = "2019-06-01"
 
         self.config = config
 
     def list(
-            self, custom_headers=None, raw=False, **operation_config):
+            self, expand_detector=None, custom_headers=None, raw=False, **operation_config):
         """List all the existing Smart Detector alert rules within the
         subscription.
 
+        :param expand_detector: Indicates if Smart Detector should be
+         expanded.
+        :type expand_detector: bool
         :param dict custom_headers: headers that will be added to the request
         :param bool raw: returns the direct response alongside the
          deserialized response
@@ -52,8 +57,7 @@ class SmartDetectorAlertRulesOperations(object):
         :raises:
          :class:`ErrorResponse1Exception<azure.mgmt.alertsmanagement.models.ErrorResponse1Exception>`
         """
-        def internal_paging(next_link=None, raw=False):
-
+        def prepare_request(next_link=None):
             if not next_link:
                 # Construct URL
                 url = self.list.metadata['url']
@@ -65,6 +69,8 @@ class SmartDetectorAlertRulesOperations(object):
                 # Construct parameters
                 query_parameters = {}
                 query_parameters['api-version'] = self._serialize.query("self.api_version", self.api_version, 'str')
+                if expand_detector is not None:
+                    query_parameters['expandDetector'] = self._serialize.query("expand_detector", expand_detector, 'bool')
 
             else:
                 url = next_link
@@ -82,6 +88,11 @@ class SmartDetectorAlertRulesOperations(object):
 
             # Construct and send request
             request = self._client.get(url, query_parameters, header_parameters)
+            return request
+
+        def internal_paging(next_link=None):
+            request = prepare_request(next_link)
+
             response = self._client.send(request, stream=False, **operation_config)
 
             if response.status_code not in [200]:
@@ -90,23 +101,24 @@ class SmartDetectorAlertRulesOperations(object):
             return response
 
         # Deserialize response
-        deserialized = models.AlertRulePaged(internal_paging, self._deserialize.dependencies)
-
+        header_dict = None
         if raw:
             header_dict = {}
-            client_raw_response = models.AlertRulePaged(internal_paging, self._deserialize.dependencies, header_dict)
-            return client_raw_response
+        deserialized = models.AlertRulePaged(internal_paging, self._deserialize.dependencies, header_dict)
 
         return deserialized
     list.metadata = {'url': '/subscriptions/{subscriptionId}/providers/microsoft.alertsManagement/smartDetectorAlertRules'}
 
     def list_by_resource_group(
-            self, resource_group_name, custom_headers=None, raw=False, **operation_config):
+            self, resource_group_name, expand_detector=None, custom_headers=None, raw=False, **operation_config):
         """List all the existing Smart Detector alert rules within the
         subscription and resource group.
 
         :param resource_group_name: The name of the resource group.
         :type resource_group_name: str
+        :param expand_detector: Indicates if Smart Detector should be
+         expanded.
+        :type expand_detector: bool
         :param dict custom_headers: headers that will be added to the request
         :param bool raw: returns the direct response alongside the
          deserialized response
@@ -118,8 +130,7 @@ class SmartDetectorAlertRulesOperations(object):
         :raises:
          :class:`ErrorResponse1Exception<azure.mgmt.alertsmanagement.models.ErrorResponse1Exception>`
         """
-        def internal_paging(next_link=None, raw=False):
-
+        def prepare_request(next_link=None):
             if not next_link:
                 # Construct URL
                 url = self.list_by_resource_group.metadata['url']
@@ -132,6 +143,8 @@ class SmartDetectorAlertRulesOperations(object):
                 # Construct parameters
                 query_parameters = {}
                 query_parameters['api-version'] = self._serialize.query("self.api_version", self.api_version, 'str')
+                if expand_detector is not None:
+                    query_parameters['expandDetector'] = self._serialize.query("expand_detector", expand_detector, 'bool')
 
             else:
                 url = next_link
@@ -149,6 +162,11 @@ class SmartDetectorAlertRulesOperations(object):
 
             # Construct and send request
             request = self._client.get(url, query_parameters, header_parameters)
+            return request
+
+        def internal_paging(next_link=None):
+            request = prepare_request(next_link)
+
             response = self._client.send(request, stream=False, **operation_config)
 
             if response.status_code not in [200]:
@@ -157,12 +175,10 @@ class SmartDetectorAlertRulesOperations(object):
             return response
 
         # Deserialize response
-        deserialized = models.AlertRulePaged(internal_paging, self._deserialize.dependencies)
-
+        header_dict = None
         if raw:
             header_dict = {}
-            client_raw_response = models.AlertRulePaged(internal_paging, self._deserialize.dependencies, header_dict)
-            return client_raw_response
+        deserialized = models.AlertRulePaged(internal_paging, self._deserialize.dependencies, header_dict)
 
         return deserialized
     list_by_resource_group.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/microsoft.alertsManagement/smartDetectorAlertRules'}
@@ -222,7 +238,6 @@ class SmartDetectorAlertRulesOperations(object):
             raise models.ErrorResponse1Exception(self._deserialize, response)
 
         deserialized = None
-
         if response.status_code == 200:
             deserialized = self._deserialize('AlertRule', response)
 
@@ -289,7 +304,6 @@ class SmartDetectorAlertRulesOperations(object):
             raise models.ErrorResponse1Exception(self._deserialize, response)
 
         deserialized = None
-
         if response.status_code == 200:
             deserialized = self._deserialize('AlertRule', response)
         if response.status_code == 201:
@@ -301,6 +315,73 @@ class SmartDetectorAlertRulesOperations(object):
 
         return deserialized
     create_or_update.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/microsoft.alertsManagement/smartDetectorAlertRules/{alertRuleName}'}
+
+    def patch(
+            self, resource_group_name, alert_rule_name, parameters, custom_headers=None, raw=False, **operation_config):
+        """Patch a specific Smart Detector alert rule.
+
+        :param resource_group_name: The name of the resource group.
+        :type resource_group_name: str
+        :param alert_rule_name: The name of the alert rule.
+        :type alert_rule_name: str
+        :param parameters: Parameters supplied to the operation.
+        :type parameters:
+         ~azure.mgmt.alertsmanagement.models.AlertRulePatchObject
+        :param dict custom_headers: headers that will be added to the request
+        :param bool raw: returns the direct response alongside the
+         deserialized response
+        :param operation_config: :ref:`Operation configuration
+         overrides<msrest:optionsforoperations>`.
+        :return: AlertRule or ClientRawResponse if raw=true
+        :rtype: ~azure.mgmt.alertsmanagement.models.AlertRule or
+         ~msrest.pipeline.ClientRawResponse
+        :raises:
+         :class:`ErrorResponse1Exception<azure.mgmt.alertsmanagement.models.ErrorResponse1Exception>`
+        """
+        # Construct URL
+        url = self.patch.metadata['url']
+        path_format_arguments = {
+            'subscriptionId': self._serialize.url("self.config.subscription_id", self.config.subscription_id, 'str'),
+            'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
+            'alertRuleName': self._serialize.url("alert_rule_name", alert_rule_name, 'str')
+        }
+        url = self._client.format_url(url, **path_format_arguments)
+
+        # Construct parameters
+        query_parameters = {}
+        query_parameters['api-version'] = self._serialize.query("self.api_version", self.api_version, 'str')
+
+        # Construct headers
+        header_parameters = {}
+        header_parameters['Accept'] = 'application/json'
+        header_parameters['Content-Type'] = 'application/json; charset=utf-8'
+        if self.config.generate_client_request_id:
+            header_parameters['x-ms-client-request-id'] = str(uuid.uuid1())
+        if custom_headers:
+            header_parameters.update(custom_headers)
+        if self.config.accept_language is not None:
+            header_parameters['accept-language'] = self._serialize.header("self.config.accept_language", self.config.accept_language, 'str')
+
+        # Construct body
+        body_content = self._serialize.body(parameters, 'AlertRulePatchObject')
+
+        # Construct and send request
+        request = self._client.patch(url, query_parameters, header_parameters, body_content)
+        response = self._client.send(request, stream=False, **operation_config)
+
+        if response.status_code not in [200]:
+            raise models.ErrorResponse1Exception(self._deserialize, response)
+
+        deserialized = None
+        if response.status_code == 200:
+            deserialized = self._deserialize('AlertRule', response)
+
+        if raw:
+            client_raw_response = ClientRawResponse(deserialized, response)
+            return client_raw_response
+
+        return deserialized
+    patch.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/microsoft.alertsManagement/smartDetectorAlertRules/{alertRuleName}'}
 
     def delete(
             self, resource_group_name, alert_rule_name, custom_headers=None, raw=False, **operation_config):