+++ sdk/alertsmanagement/azure-mgmt-alertsmanagement/HISTORY.rst
@@ -3,6 +3,15 @@
 Release History
 ===============
 
+0.1.1 (2019-07-20)
+++++++++++++++++++
+
+**Features**
+
+- Added operation AlertsOperations.meta_data
+- Added operation group SmartDetectorAlertRulesOperations
+- Added operation group ActionRulesOperations
+
 0.1.0 (2018-09-17)
 ++++++++++++++++++